requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
aiosqlite==0.19.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiogram==3.3.0
python-dotenv==1.0.1
alembic==1.13.1
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
python-telegram-bot
flask
requests 